+++ index.js
@@ -1,2 +1,4 @@
+/* eslint-disable unicorn/no-empty-file */
+
 // This file is left here, intentionally blank, until I find a better way to
 // export 'PMEUserConfig'.